moved supervisor to discovery
